feat(wasm): resctrict serviceWorker scope
