+++ examples/utooweb-demo/index.html
@@ -8,8 +8,6 @@
 </head>
 
 <body>
-  <h1>Check the console!</h1>
-  <p>This is a minimal utooweb example.</p>
   <div id="root">
 </body>
