+++ packages/utoo-web/src/serviceWorker.ts
@@ -3,37 +3,36 @@ import { ServiceWorkerHandShake } from "./message";
 
 declare let self: ServiceWorkerGlobalScope;
 
-let __resolve__: () => void;
+let _resolve: () => void;
 
-let _promise__: Promise<void> = new Promise((resolve) => {
-  __resolve__ = resolve;
+let _promise: Promise<void> = new Promise((resolve) => {
+  _resolve = resolve;
 });
 
-let projectEndpoint: ProjectEndpoint | undefined;
+let _projectEndpoint: ProjectEndpoint;
 
-let proxiedResourcePath: string | undefined;
+let _serviceWorkerScope: string;
 
 self.addEventListener("message", (event) => {
   if (event.data && event.data[ServiceWorkerHandShake] === true) {
-    proxiedResourcePath = event.data.proxiedResourcePath;
-    projectEndpoint = Project.fork(
+    _serviceWorkerScope = event.data.scope;
+    _projectEndpoint = Project.fork(
       new MessageChannel(),
       event.source as Client,
     );
-    __resolve__();
+    _resolve();
   }
 });
 
 self.addEventListener("fetch", async (event: FetchEvent) => {
   const { url, referrer } = event.request;
   if (
-    new URL(url).pathname.startsWith(`/${proxiedResourcePath}`) ||
-    (referrer &&
-      new URL(referrer).pathname.startsWith(`/${proxiedResourcePath}`))
+    new URL(url).pathname.startsWith(_serviceWorkerScope) ||
+    (referrer && new URL(referrer).pathname.startsWith(_serviceWorkerScope))
   ) {
-    await _promise__;
+    await _promise;
     const projectPath =
-      "." + new URL(url).pathname.replace(`/${proxiedResourcePath}`, "");
+      "." + new URL(url).pathname.replace(_serviceWorkerScope, "");
     event.respondWith(readFileFromProject(projectPath));
   } else {
     event.respondWith(fetch(event.request));
@@ -42,7 +41,7 @@ self.addEventListener("fetch", async (event: FetchEvent) => {
 
 async function readFileFromProject(projectPath: string): Promise<Response> {
   try {
-    const content = await projectEndpoint!.readFile(projectPath);
+    const content = await _projectEndpoint.readFile(projectPath);
 
     let mimeType = "application/octet-stream";
     if (projectPath.endsWith(".js")) {
